Automatically download the shared libraries during composer install
