+++ composer.json
@@ -20,7 +20,8 @@
     "codewithkyrian/jinja-php": "^1.0",
     "symfony/console": "^6.4|^7.0",
     "imagine/imagine": "^1.3",
-    "rokka/imagine-vips": "^0.31.0"
+    "rokka/imagine-vips": "^0.31.0",
+    "codewithkyrian/onnxruntime-downloader-plugin": "^1.0"
   },
   "require-dev": {
     "pestphp/pest": "^2.31",
@@ -44,7 +45,8 @@
   ],
   "config": {
     "allow-plugins": {
-      "pestphp/pest-plugin": true
+      "pestphp/pest-plugin": true,
+      "codewithkyrian/onnxruntime-downloader-plugin": true
     }
   },
   "bin": [
